Clean contact list (#250)
* WIP

* WIP 2

* SEperated contact list hooks from resources, seperating concerns
