+++ packages/esm-patient-clinical-view-app/src/contact-list/contact-actions.component.tsx
@@ -4,11 +4,9 @@ import { useConfig } from '@openmrs/esm-framework';
 import { launchPatientWorkspace } from '@openmrs/esm-patient-common-lib';
 import React from 'react';
 import { ConfigObject } from '../config-schema';
-import {
-  getHivStatusBasedOnEnrollmentAndHTSEncounters,
-  useRelativeHTSEncounter,
-  useRelativeHivEnrollment,
-} from './contact-list.resource';
+import useRelativeHTSEncounter from '../hooks/useRelativeHTSEncounter';
+import useRelativeHivEnrollment from '../hooks/useRelativeHivEnrollment';
+import { getHivStatusBasedOnEnrollmentAndHTSEncounters } from './contact-list.resource';
 
 interface ContactActionsProps {
   relativeUuid: string;
