+++ packages/esm-patient-clinical-view-app/src/contact-list/hiv-status.component.tsx
@@ -1,10 +1,8 @@
 import { SkeletonText } from '@carbon/react';
 import React from 'react';
-import {
-  getHivStatusBasedOnEnrollmentAndHTSEncounters,
-  useRelativeHTSEncounter,
-  useRelativeHivEnrollment,
-} from './contact-list.resource';
+import useRelativeHTSEncounter from '../hooks/useRelativeHTSEncounter';
+import useRelativeHivEnrollment from '../hooks/useRelativeHivEnrollment';
+import { getHivStatusBasedOnEnrollmentAndHTSEncounters } from './contact-list.resource';
 
 interface HIVStatusProps {
   relativeUuid: string;
